// projeto final/cinematch-api/src/movies/dto/create-movie.dto.ts
export class CreateMovieDto {
  seriesTitle?: string;
  genre?: string;
  director?: string;
  runtime?: string;
  releasedYear?: number;
  posterLink?: string;
  certificate?: string;
  imdbRating?: number;
  overview?: string;
  metaScore?: number;
  star1?: string;
  star2?: string;
  star3?: string;
  star4?: string;
  noOfVotes?: number;
  gross?: string;
}